Move clocksource_default_clock to regular segment
This patch removes the __init specifier of all occurences of the
function, thus moving it to the normal kernel code segment instead
of the discardable __init one. It is necessary for invoking the
function some time after system boot, as will be needed for the
SCC-specific TSC clocksource patch to track cpufreq changes.