SCC: Fix lapic_clockevent parameters
The manipulated LAPIC clock calibration code for the SCC uses the
pre-configured busclock to calculate a calibration_result, but it
does not fill in other fields of the lapic_clockevent structure
the regular calibration path would have done. This patch attempts
to fix it, by first calculating a "delta" value just like the normal
code would have done, then deriving the parameters for the clock device.